use unsigned int in as_uint

+++ src/rlso/so-as.c
@@ -208,7 +208,7 @@ ErrDecl so_as_ushort(So so, unsigned short *out, int base) {
 
 ErrDecl so_as_uint(So so, unsigned int *out, int base) {
     so_as_get_base(&so, &base);
-    SO_AS_WITH_BASE(unsigned long, so, base, false, out, UINT_MAX);
+    SO_AS_WITH_BASE(unsigned int, so, base, false, out, UINT_MAX);
 }
 
 ErrDecl so_as_ulong(So so, unsigned long *out, int base) {
